+++ hack/jenkins/cron/cleanup_and_reboot_Linux.sh
@@ -22,7 +22,7 @@ function check_jenkins() {
   fi
   pstree "${jenkins_pid}" \
         | egrep -i 'bash|integration|e2e|minikube' \
-        && echo "tests are is running on pid ${jenkins_pid} ..." \
+        && echo "tests are running on pid ${jenkins_pid} ..." \
         && exit 1
 }