+++ hack/jenkins/test-flake-chart/report_flakes.sh
@@ -85,11 +85,7 @@ awk -F, 'NR>1 {
   | sort -g -t, -k2,2 \
   >> "$TMP_FAILED_RATES"
 
-# Filter out arm64, crio, and QEMU tests until they're more stable
-TMP_FAILED_RATES_FILTERED=$(mktemp)
-grep -v "arm64\|crio\|QEMU" "$TMP_FAILED_RATES" > "$TMP_FAILED_RATES_FILTERED"
-
-FAILED_RATES_LINES=$(wc -l < "$TMP_FAILED_RATES_FILTERED")
+FAILED_RATES_LINES=$(wc -l < "$TMP_FAILED_RATES")
 if [[ "$FAILED_RATES_LINES" -eq 0 ]]; then
   echo "No failed tests! Aborting without commenting..." 1>&2
   exit 0
@@ -106,7 +102,7 @@ TEST_GOPOGH_LINK_FORMAT='https://storage.googleapis.com/minikube-builds/logs/'${
 # 1) Get the first $MAX_REPORTED_TESTS lines.
 # 2) Print a row in the table with the environment, test name, flake rate, and a link to the flake chart for that test.
 # 3) Append these rows to file $TMP_COMMENT.
-head -n "$MAX_REPORTED_TESTS" "$TMP_FAILED_RATES_FILTERED" \
+head -n "$MAX_REPORTED_TESTS" "$TMP_FAILED_RATES" \
   | awk '-F[:,]' '{
       if ($3 != "n/a") {
         rate_text = sprintf("%3$s ([chart]('$TEST_CHART_LINK_FORMAT'))", $1, $2, $3)