+++ hack/update/get_version/get_version.go
@@ -30,6 +30,7 @@ type dependency struct {
 var dependencies = map[string]dependency{
 	"buildkit":       {"deploy/iso/minikube-iso/arch/x86_64/package/buildkit-bin/buildkit-bin.mk", `BUILDKIT_BIN_VERSION = (.*)`},
 	"cloud-spanner":  {"pkg/minikube/assets/addons.go", `cloud-spanner-emulator/emulator:(.*)@`},
+	"cni-plugins":    {"deploy/iso/minikube-iso/arch/x86_64/package/cni-plugins/cni-plugins.mk", `CNI_PLUGINS_VERSION = (.*)`},
 	"containerd":     {"deploy/iso/minikube-iso/arch/x86_64/package/containerd-bin/containerd-bin.mk", `CONTAINERD_BIN_VERSION = (.*)`},
 	"cri-o":          {"deploy/iso/minikube-iso/package/crio-bin/crio-bin.mk", `CRIO_BIN_VERSION = (.*)`},
 	"docker":         {"deploy/iso/minikube-iso/arch/x86_64/package/docker-bin/docker-bin.mk", `DOCKER_BIN_VERSION = (.*)`},
@@ -41,6 +42,7 @@ var dependencies = map[string]dependency{
 	"hugo":           {"netlify.toml", `HUGO_VERSION = "(.*)"`},
 	"metrics-server": {"pkg/minikube/assets/addons.go", `metrics-server/metrics-server:(.*)@`},
 	"runc":           {"deploy/iso/minikube-iso/package/runc-master/runc-master.mk", `RUNC_MASTER_VERSION = (.*)`},
+	"ubuntu":         {"deploy/kicbase/Dockerfile", `UBUNTU_FOCAL_IMAGE="(.*)"`},
 }
 
 func main() {